requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
uuid==1.30
